requirements.txt
Faker
Pandas
Numpy
google-cloud-pubsub
google-cloud-bigquery
Findspark
pyspark
jupyter